madgraph: added madgraph alias, gfortran dependency
To avoid confusion with aMC@NLO (which is a beta version of MadGraph
which I have decided to give its own formula), I create a symlink to
`madgraph`. I add a gfortran dependency because this is required to
actually run MadGraph (note: this now fails brew audit, but that
warning is intended for gfortran dependence during the build). Finally,
I use `ditto` rather than `cp -r` since it is a bit nicer.