rivet: provide explicit Boost path
This should fix a memmory leak in Rivet cut shared pointers that was
observed on Linux system. The bug probably happens only when building an
out of tree analysis. Without an explicit path for Boost, the effective
Boost path would not propagate into the rivet-buildanalysis tool so that
analysis could have been built with a different Boost headers than the
ones that were used for the Rivet.