lhapdf: add patch to fix install subcommand
... which is broken since hepforge was upgraded a few months ago.  The
underlying issue might be related to redirections when trying to
download the PDF set. Anyway, until this is fixed upstream our patch
fixes the issue (although we do not have a nice download status bar).
This is particularly important because e.g. the herwig formula depends
on lhapdf-install to work.